Changes to be committed: 	modified:   package.json 	modified:   src/imports/config-module.import.ts 	modified:   src/utils/get-app-config-for-external-use.util.spec.ts 	modified:   src/utils/index.ts 	renamed:    src/utils/registerValidatedConfig.util.spec.ts -> src/utils/register-validated-config.util.spec.ts 	renamed:    src/utils/registerValidatedConfig.util.ts -> src/utils/register-validated-config.util.ts
