+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nest-base",
-  "version": "1.3.0",
+  "version": "1.4.0",
   "description": "",
   "license": "MIT",
   "keywords": [],
