+++ src/imports/config-module.import.ts
@@ -1,7 +1,7 @@
 import { ConfigModule } from '@nestjs/config';
 
 import { AppConfig } from '../config';
-import { registerValidatedConfig } from '../utils';
+import { registerValidatedConfig } from '../utils/register-validated-config.util';
 
 export class AppConfigModule {
   static forRoot() {
