+++ src/utils/index.ts
@@ -2,6 +2,6 @@ export * from './byte-length-humanize.util';
 export * from './get-app-config-for-external-use.util';
 export * from './get-package-metadata.util';
 export * from './read-json-file';
-export * from './registerValidatedConfig.util';
+export * from './register-validated-config.util';
 export * from './string.util';
 export * from './winston-logger.util';
