+++ src/utils/register-validated-config.util.ts
